代码生成

+++ keiskei-framework-core/src/main/java/top/keiskeiframework/common/util/json/MoneyDeserializer.java
@@ -9,12 +9,12 @@ import com.fasterxml.jackson.databind.util.StdConverter;
  * @author 陈加敏 right_way@foxmail.com
  * @since 2019/8/5 20:50
  */
-public class MoneyDeserializer extends StdConverter<Double, Integer> {
+public class MoneyDeserializer extends StdConverter<Double, Long> {
 
 
     @Override
-    public Integer convert(Double val) {
-        return (int) (val * 100);
+    public Long convert(Double val) {
+        return (long) (val * 100);
     }
